+++ src/assets/themes/index.ts
@@ -1,14 +1,4 @@
 import { createTheme } from "@mui/material";
-import type { PaletteColor, PaletteColorOptions } from "@mui/material";
-
-declare module "@mui/material/styles" {
-  interface Palette {
-    accent: PaletteColor;
-  }
-  interface PaletteOptions {
-    accent?: PaletteColorOptions;
-  }
-}
 
 // ここを変更したらvariable.scssも変更する
 const COLOR_PALETTE = {
