+++ src/libs/types/mui.d.ts
@@ -1,0 +1,10 @@
+import type { PaletteColor, PaletteColorOptions } from "@mui/material";
+
+declare module "@mui/material/styles" {
+  interface Palette {
+    accent: PaletteColor;
+  }
+  interface PaletteOptions {
+    accent?: PaletteColorOptions;
+  }
+}
